Update method of removing ads as old method was not working anymore
The old method based on the CSS attribute .ad-indicator is not working anymore. Instead we can catch a difference between ads and non ads items with an attribute where ads are indicated in the place of the shop/seller name:  .logged . Items with ads have this attribute: .logged.listing-link while non ads have only .listing-link 
So we can just catch it and delete its parents and voila. 
I also tried while connected, because as it is called logged, I was a bit afraid that some other things will be broken, but it does not seem so. I did not test the whole site though so it might be on some other page (I hope not).